change some migrations

+++ app/database/migrations/2019_08_21_013700_create_user_invoices.php
@@ -17,7 +17,7 @@ class CreateUserInvoices extends Migration
             $table->timestamps();
             $table->integer('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
-            $table->integer('cents');
+            $table->float('cents');
             $table->string('source');
             $table->string('status');
             $table->integer('balance');
