+++ app/database/migrations/2019_12_01_234439_add_size_to_recordings.php
@@ -14,7 +14,7 @@ class AddSizeToRecordings extends Migration
     {
         Schema::table('recordings', function (Blueprint $table) {
             //
-            $table->integer('size')->default(0);
+            $table->float('size')->default(0.0);
         });
     }
